move common functionality into `common.dart`
